+++ files/ptu_heart_beat_timeout_ms_app_1234567.json
@@ -2317,7 +2317,7 @@
         "steal_focus": false,
         "priority": "NONE",
         "default_hmi": "NONE",
-	"heart_beat_timeout_ms":2000,
+	"heart_beat_timeout_ms":4000,
         "groups": [
           "BaseBeforeDataConsent"
         ]
